+++ src/main/java/com/herokuapp/entity/Khachvanglai.java
@@ -31,7 +31,7 @@ public class Khachvanglai implements Serializable {
 	private String ho;
 
 	@Column(nullable=false)
-	private int sdt;
+	private Long sdt;
 
 	@Column(nullable=false, length=30)
 	private String ten;
@@ -84,11 +84,11 @@ public class Khachvanglai implements Serializable {
 		this.ho = ho;
 	}
 
-	public int getSdt() {
+	public Long getSdt() {
 		return this.sdt;
 	}
 
-	public void setSdt(int sdt) {
+	public void setSdt(Long sdt) {
 		this.sdt = sdt;
 	}
